crawl_scheduler/requirements.txt
cloud-sql-python-connector[pg8000]==0.4.1
google-cloud-pubsub==2.8.0
python-dotenv==0.19.1
psycopg2-binary==2.8.6